+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Fixed
 
+* Fixed: `C-w _` (set current group height as high as possible) doesn't always work correctly
+* Fixed: `C-w |` (set current group width as wide as possible) doesn't always work correctly
+* Fixed: `C-w =` (resize all groups equally) doesn't always work correctly
 * Fixed: `gJ`
 * Fixed: `gx` should ignore trailing full stops
 * Fixed: `gx` doesn't work on markdown links